backend/requirements.txt
django==2.0.5
djangorestframework==3.8.2
graphene-django==2.2.0
django-graphql-jwt==0.1.10
django-rq==1.2.0
requests==2.19.1
redis==2.10.6

# Django-cities (cities)
swapper==1.1.0
django-model-utils==3.1.2
tqdm==4.24.0

# Testing
Faker==0.8.15
factory-boy==2.11.1
snapshottest==0.5.0
